moved webapp code to separate module
