+++ backtrader_plotting/bokeh/__init__.py
@@ -1,0 +1,1 @@
+from backtrader_plotting.bokeh.bokeh_webapp import BokehWebapp
